Allow for other content types other than application/json
Rails can handle multiple content types like application/json,
multipart/form-data. It parses them and builds the ActionController::
Parameters with the params that includes the path_parameters and
query_parameters. During a post call where we try to get just the
body parameters by reparsing the body, if we just exclude the path
parameters from the params we wont have to reparse the request.body.
Also the correct content-type needs to be sent to the openapi-parser.